pyproject.toml
[tool.poetry]
name = "conv_2020_website"
version = "0.2.0"
description = ""
authors = ["Kim van Wyk <kimv@sahomeloans.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-dateutil = "^2.8"
pyperclip = "^1.7"
awscli = "^1.16"
sqlalchemy = "^1.3"
psycopg2-binary = "^2.8"
matplotlib = "^3.1.2"
pandas = "^0.25.3"
black = "^19.10b0"
requests = "^2.23.0"
attrs = "^19.3.0"
dateparser = "^0.7.6"

[tool.poetry.dev-dependencies]
ptpython = "^2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
